+++ ProdManager/models/Scope.py
@@ -56,4 +56,4 @@ class Scope(db.Model):
     return cls.name.asc()
 
   def monitors_count(self, serialize=False):
-    return Monitor.count_by_status(self.monitors, serialize)
+    return Monitor.count_by_status(serialize, (Monitor.scope_id == self.id,))